data/generators/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
pathlib>=1.0.1
typing>=3.7.4.3 